# api-spec/swagger.yaml
---
swagger: "2.0"
info:
  description: API do sistema de Licenciamento de Ambulantes de Salvador
  version: 1.1.0
  title: LAS API
  contact:
    email: admin@las.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /tiagolpadua/las-api/1.1.0
tags:
- name: usuario
  description: Operações com usuários
- name: evento
  description: Operações com eventos
- name: tipo-venda
  description: Operações com tipos de vendas
- name: dados-pessoais
  description: Operações com dados pessoais
- name: contatos
  description: Operações com contato
- name: senha
  description: Operações com senha
- name: endereco
  description: Operações com endereço
- name: tipo-venda-ambulante
  description: Operações com tipo de venda do ambulante
- name: documento
  description: Operações com documento do usuário
schemes:
- https
paths:
  /usuarios:
    get:
      tags:
      - usuario
      summary: Lista os usuarios
      description: Lista os usuários do LAS-API
      operationId: listarUsuarios
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Resultado da listagem
          schema:
            type: array
            items:
              $ref: '#/definitions/Usuario'
    post:
      tags:
      - usuario
      summary: Inclui um usuário
      description: Inclui um usuário no LAS-API
      operationId: incluirUsuario
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: usuario
        description: Usuário a ser incluído
        required: false
        schema:
          $ref: '#/definitions/Usuario'
      responses:
        "201":
          description: Usuário incluído com sucesso
        "400":
          description: Entrada inválida
  /usuarios/{usuarioId}:
    get:
      tags:
      - usuario
      summary: Busca um usuário pelo ID
      description: Busca um único usuário
      operationId: buscarUsuarioPorId
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser retornado
        required: true
        type: integer
      responses:
        "200":
          description: Operação bem sucedida
          schema:
            $ref: '#/definitions/Usuario'
        "400":
          description: Id inválido fornecido
        "404":
          description: Usuário não encontrado
    put:
      tags:
      - usuario
      summary: Atualiza um usuário no LAS-API com os dados do formulário
      operationId: atualizarUsuario
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser atualizado
        required: true
        type: integer
      - in: body
        name: usuario
        description: Usuário a ser atualizado
        required: false
        schema:
          $ref: '#/definitions/Usuario'
      responses:
        "204":
          description: Usuário atualizado com sucesso
        "405":
          description: Entrada inválida
    delete:
      tags:
      - usuario
      summary: Exclui um usuário
      operationId: excluirUsuario
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser excluído
        required: true
        type: integer
      responses:
        "204":
          description: Usuário excluído com sucesso
        "400":
          description: Id inválido fornecido
        "404":
          description: Usuário não encontrado
  /usuarios/{usuarioId}/dados-pessoais:
    get:
      tags:
      - dados-pessoais
      summary: Busca os dados pessoais do usuário
      operationId: buscarDadosPessoaisDoUsuario
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a retornar os dados pessoais
        required: true
        type: integer
      responses:
        "200":
          description: Operação bem sucedida
          schema:
            $ref: '#/definitions/DadosPessoais'
        "400":
          description: Id inválido fornecido
        "404":
          description: Usuário não encontrado
    put:
      tags:
      - dados-pessoais
      summary: Atualiza os dados pessoais do usuário
      operationId: atualizarDadosPessoaisDoUsuario
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser atualizado
        required: true
        type: integer
      - in: body
        name: dadosPessoais
        description: Dados pessoais do usuário a ser atualizado
        required: false
        schema:
          $ref: '#/definitions/DadosPessoais'
      responses:
        "204":
          description: Dados pessoais atualizados com sucesso
        "405":
          description: Entrada inválida
  /usuarios/{usuarioId}/contatos:
    get:
      tags:
      - contatos
      summary: Busca os contatos do usuário
      operationId: buscarContatosDoUsuario
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a retornar os contatos
        required: true
        type: integer
      responses:
        "200":
          description: Operação bem sucedida
          schema:
            $ref: '#/definitions/Contatos'
        "400":
          description: Id inválido fornecido
        "404":
          description: Usuário não encontrado
    put:
      tags:
      - contatos
      summary: Atualiza os contatos do usuário
      operationId: atualizarContatosDoUsuario
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser atualizado
        required: true
        type: integer
      - in: body
        name: contatos
        description: Contatos do usuário a ser atualizado
        required: false
        schema:
          $ref: '#/definitions/Contatos'
      responses:
        "204":
          description: Contatos atualizados com sucesso
        "405":
          description: Entrada inválida
  /usuarios/{usuarioId}/endereco:
    get:
      tags:
      - endereco
      summary: Busca o endereço do usuário
      operationId: buscarEndereçoDoUsuario
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a retornar o endereço
        required: true
        type: integer
      responses:
        "200":
          description: Operação bem sucedida
          schema:
            $ref: '#/definitions/Endereco'
        "400":
          description: Id inválido fornecido
        "404":
          description: Usuário não encontrado
    put:
      tags:
      - contatos
      summary: Atualiza o endereço do usuário
      operationId: atualizarEnderecoDoUsuario
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser atualizado
        required: true
        type: integer
      - in: body
        name: contatos
        description: Endereço do usuário a ser atualizado
        required: false
        schema:
          $ref: '#/definitions/Endereco'
      responses:
        "204":
          description: Endereço atualizado com sucesso
        "405":
          description: Entrada inválida
  /usuarios/{usuarioId}/documentos/{tipoDocumento}:
    get:
      tags:
      - documento
      summary: Retorna um documento do usuário
      operationId: buscarDocumentoDoUsuario
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a retornar o tipo de venda
        required: true
        type: integer
      - name: tipoDocumento
        in: path
        description: Tipo do documento a ser retornado
        required: true
        type: string
      responses:
        "200":
          description: Operação bem sucedida
          schema:
            $ref: '#/definitions/Documento'
        "400":
          description: Id inválido fornecido
        "404":
          description: Usuário não encontrado
    put:
      tags:
      - documento
      summary: Atualiza um documento do usuário
      operationId: atualizarDocumentoDoUsuario
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser atualizado
        required: true
        type: integer
      - name: tipoDocumento
        in: path
        description: Tipo do documento a ser atualizado
        required: true
        type: string
      - in: body
        name: documento
        description: Documento a ser atualizado
        required: false
        schema:
          $ref: '#/definitions/Documento'
      responses:
        "204":
          description: Documento atualizado com sucesso
        "405":
          description: Entrada inválida
  /usuarios/{usuarioId}/senha:
    put:
      tags:
      - senha
      summary: Atualiza a senha do usuário
      operationId: atualizarSenhaDoUsuario
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: usuarioId
        in: path
        description: Id do usuário a ser atualizado
        required: true
        type: integer
      - in: body
        name: senha
        description: Senha do usuário a ser atualizado
        required: false
        schema:
          $ref: '#/definitions/Senha'
      responses:
        "204":
          description: Senha atualizada com sucesso
        "405":
          description: Entrada inválida
  /eventos:
    get:
      tags:
      - evento
      summary: Lista os eventos
      description: Lista os eventos do LAS-API com a descrição truncada em 100 caracteres
      operationId: listarEventos
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Resultado da listagem
          schema:
            type: array
            items:
              $ref: '#/definitions/Evento'
    post:
      tags:
      - evento
      summary: Inclui um evento
      description: Inclui um evento no LAS-API
      operationId: incluirEvento
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: evento
        description: Evento a ser incluído
        required: false
        schema:
          $ref: '#/definitions/Evento'
      responses:
        "201":
          description: Evento incluído com sucesso
        "400":
          description: Entrada inválida
  /eventos/status/{statusEvento}:
    get:
      tags:
      - evento
      summary: Busca os eventos pelo staus
      description: Busca eventos de acordo com o status informado
      operationId: buscarEventoPorStatus
      produces:
      - application/json
      parameters:
      - name: statusEvento
        in: path
        description: Status dos eventos a serem listados
        required: true
        type: string
      responses:
        "200":
          description: Resultado da listagem
          schema:
            type: array
            items:
              $ref: '#/definitions/Evento'
        "400":
          description: Status inválido fornecido
        "404":
          description: Evento não encontrado
  /eventos/{eventoId}:
    get:
      tags:
      - evento
      summary: Busca um evento pelo id
      description: Busca um único evento
      operationId: buscarEventoPorId
      produces:
      - application/json
      parameters:
      - name: eventoId
        in: path
        description: Id do evento a ser retornado
        required: true
        type: integer
      responses:
        "200":
          description: Operação bem sucedida
          schema:
            $ref: '#/definitions/Evento'
        "400":
          description: Id inválido fornecido
        "404":
          description: Evento não encontrado
    put:
      tags:
      - evento
      summary: Atualiza um evento no LAS-API com os dados enviados
      operationId: atualizarEvento
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: eventoId
        in: path
        description: Id do evento a ser atualizado
        required: true
        type: integer
      - in: body
        name: evento
        description: Evento a ser atualizado
        required: false
        schema:
          $ref: '#/definitions/Evento'
      responses:
        "204":
          description: Evento atualizado com sucesso
        "405":
          description: Entrada inválida
    delete:
      tags:
      - evento
      summary: Exclui um evento
      operationId: excluirEvento
      produces:
      - application/json
      parameters:
      - name: eventoId
        in: path
        description: Id do evento a ser excluído
        required: true
        type: integer
      responses:
        "204":
          description: Evento excluído com sucesso
        "400":
          description: Id inválido fornecido
        "404":
          description: Evento não encontrado
  /ufs:
    get:
      tags:
      - endereco
      summary: Lista as UFs
      operationId: listarUFs
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Resultado da listagem
          schema:
            type: array
            items:
              type: string
            example:
            - BA
            - SP
            - DF
  /ufs/{uf}/municipios:
    get:
      tags:
      - endereco
      summary: Lista os municípios da UF
      operationId: listarMunicipiosUF
      produces:
      - application/json
      parameters:
      - name: uf
        in: path
        description: UF para listar os municípios
        required: true
        type: string
      responses:
        "200":
          description: Resultado da listagem
          schema:
            type: array
            items:
              type: string
            example:
            - Salvador
            - Ilhéus
            - Brasília
            - São Paulo
  /tipos-venda:
    get:
      tags:
      - tipo-venda
      summary: Lista os tipos de vendas
      description: |
        Lista os tipos de vendas do LAS-API
      operationId: listarTiposVendas
      produces:
      - application/json
      parameters: []
      responses:
        "200":
          description: Resultado da listagem
          schema:
            type: array
            items:
              $ref: '#/definitions/TipoVenda'
    post:
      tags:
      - tipo-venda
      summary: Inclui um tipo de venda
      description: Inclui um tipo de venda no LAS-API
      operationId: incluirTipoVenda
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: tipoVenda
        description: Tipo de Venda a ser incluído
        required: false
        schema:
          $ref: '#/definitions/TipoVenda'
      responses:
        "201":
          description: Tipo de venda incluída com sucesso
        "400":
          description: Entrada inválida
  /tipos-venda/{tipoVendaId}:
    put:
      tags:
      - tipo-venda
      summary: Atualiza um tipo de venda no LAS-API com os dados informados
      operationId: atualizarTipoVenda
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: tipoVendaId
        in: path
        description: Id do tipo de venda a ser atualizado
        required: true
        type: integer
      - in: body
        name: tipoVenda
        description: Tipo de venda a ser atualizada
        required: false
        schema:
          $ref: '#/definitions/TipoVenda'
      responses:
        "405":
          description: Entrada inválida
    delete:
      tags:
      - tipo-venda
      summary: Exclui um tipo de venda
      operationId: excluirTipoVenda
      produces:
      - application/json
      parameters:
      - name: tipoVendaId
        in: path
        description: Id do tipo de venda a ser excluída
        required: true
        type: integer
      responses:
        "204":
          description: Tipo de venda excluído com sucesso
        "400":
          description: Id inválido fornecido
        "404":
          description: Tipo de venda não encontrada
definitions:
  Usuario:
    type: object
    required:
    - id
    - nome
    - urlFotoPerfil
    properties:
      id:
        type: integer
        format: int32
        example: 12345
      nome:
        type: string
        example: Paulo Silva
        minLength: 5
      urlFotoPerfil:
        type: string
        format: uri
        example: https://randomuser.me/api/portraits/men/91.jpg
  Evento:
    type: object
    required:
    - dataFim
    - dataInicio
    - descricao
    - id
    - nome
    - urlFoto
    properties:
      id:
        type: integer
        format: int32
        example: 12345
      nome:
        type: string
        example: Carnaval 2022
        minLength: 5
      descricao:
        type: string
        example: O carnaval mais animado do Brasil
        minLength: 5
      urlFoto:
        type: string
        format: uri
        example: https://upload.wikimedia.org/wikipedia/commons/thumb/d/d3/Bloco_da_capoeira_circuito_Campo_Grande_Salvador.jpg/1920px-Bloco_da_capoeira_circuito_Campo_Grande_Salvador.jpg
      dataInicio:
        type: string
        format: date
        example: 2017-07-21T00:00:00.000+0000
      dataFim:
        type: string
        format: date
        example: 2017-07-21T00:00:00.000+0000
      status:
        type: string
        enum:
        - agendado
        - em-andamento
        - finalizado
  TipoVenda:
    type: object
    required:
    - descricao
    - id
    properties:
      id:
        type: integer
        format: int32
        example: 12345
      descricao:
        type: string
        example: Bebida
        minLength: 5
  DadosPessoais:
    type: object
    required:
    - cpf
    - dataNascimento
    - nomeCompleto
    - rg
    properties:
      nomeCompleto:
        type: string
        example: Fulano de Tal
        minLength: 5
      dataNascimento:
        type: string
        format: date
        example: 2017-07-21T00:00:00.000+0000
      rg:
        type: string
        example: 123321 SSP BA
        minLength: 5
      cpf:
        type: string
        example: "12312312312"
  Endereco:
    type: object
    required:
    - bairro
    - cep
    - endereco
    - numero
    properties:
      cep:
        type: string
        example: "72980000"
        minLength: 8
      endereco:
        type: string
        example: Rua 123
      numero:
        type: integer
        example: 23
      complemento:
        type: string
        example: Apartamento 509
      bairro:
        type: string
        example: Zona Norte
  Contatos:
    type: object
    required:
    - celular
    - email
    - telefone
    properties:
      telefone:
        type: string
        example: "6233311212"
      celular:
        type: string
        example: "62998757575"
      email:
        type: string
        format: email
        example: foobar@gmail.com
        minLength: 5
  Senha:
    type: object
    required:
    - senha
    properties:
      senha:
        type: string
        example: admin
        minLength: 5
  Documento:
    type: object
    required:
    - imagemBase64
    properties:
      imagemBase64:
        type: string
        example: iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg==